Split THTensorMath into Math and MoreMath
